requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
anthropic>=0.85.0
youtube-transcript-api>=1.2.0
yt-dlp>=2026.3.0
python-docx>=1.0.0
